added cast to float lines 308 309

+++ vcg/complex/algorithms/update/color.h
@@ -304,8 +304,8 @@ static void FaceQualityRamp(UpdateMeshType &m,bool selected=false)
   for(fi=m.face.begin();fi!=m.face.end();++fi) if(!(*fi).IsD())
   if(!selected || (*fi).IsS())
     {
-    minq=std::min(minq,(*fi).Q());
-    maxq=std::max(maxq,(*fi).Q());
+    minq=std::min(minq,(float)(*fi).Q());
+    maxq=std::max(maxq,(float)(*fi).Q());
 		}
 
   FaceQualityRamp(m,minq,maxq,selected);
